# swagger/user.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Eventol auth
  description: Eventol auth
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths
basePath: /api
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /signup:
    x-swagger-router-controller: signup
    post:
      operationId: signUp
      parameters:
          - name: user
            in: body
            description: The properties of the user
            required: true
            schema:
              $ref: '#/definitions/UserProperties'
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/JwtResponse'
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /login:
    x-swagger-router-controller: login
    post:
      operationId: logIn
      parameters:
        - name: credentials
          in: body
          description: The credentials of the user
          required: true
          schema:
            $ref: '#/definitions/Credentials'
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/JwtResponse'
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /link/appusers/{userId}:
    x-swagger-router-controller: social-login
    post:
      operationId: linkProfile
      parameters:
        - name: userId
          in: path
          description: The id of the app user
          required: true
          type: string
        - name: identity
          in: body
          description: The identity to link
          required: true
          schema:
            $ref: '#/definitions/Identity'
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/JwtResponse'
  /unlink/appusers/{userId}:
    x-swagger-router-controller: social-login
    post:
      operationId: unlink
      parameters:
        - name: userId
          in: path
          description: The id of the app user
          required: true
          type: string
        - name: identity
          in: body
          description: The identity to link
          required: true
          schema:
            $ref: '#/definitions/Identity'
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/JwtResponse'
  /app/{appId}/{provider}/login:
    x-swagger-router-controller: social-login
    get:
      operationId: socialLogin
      parameters:
        - name: appId
          in: path
          description: The clientId of the app
          required: true
          type: string
        - name: provider
          in: path
          description: The social provider
          required: true
          type: string
        - name: redirectUrl
          in: query
          required: true
          type: string
      responses:
          302:
            description: "302 response"
            headers:
              Location:
                type: "string"
  /app/{appId}/{provider}/callback:
    x-swagger-router-controller: social-login
    get:
      operationId: callback
      parameters:
        - name: appId
          in: path
          description: The clientId of the app
          required: true
          type: string
        - name: provider
          in: path
          description: The social provider
          required: true
          type: string
      responses:
        302:
          description: "302 response"
          headers:
            Location:
              type: "string"
definitions:
  Credentials:
      properties:
        username:
          type: string
        password:
          type: string
  UserProperties:
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
  JwtResponse:
    required:
      - jwt
    properties:
      jwt:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Identity:
    properties:
      provider:
        type: string
      connection:
        type: string
      jwt:
        type: string